pyproject.toml
[project]
name = "upscaler"
version = "0.1.0"
description = ""
authors = [
    {name = "Maximusin",email = "maxim.maxim.talankin@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.15"
dependencies = [
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
    "numpy (<2.3.0)",
    "pyside6 (>=6.10.1,<7.0.0)",
    "qt-material (>=2.17,<3.0)",
    "pyinstaller (>=6.17.0,<7.0.0)",
    "onnx (>=1.20.0,<2.0.0)",
    "onnxconverter-common (>=1.16.0,<2.0.0)",
    "onnxruntime-gpu[cuda,cudnn] (>=1.23.2,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
